Add - Category Repository and Dao for category added

+++ data/src/main/java/com/ajijul/elnaz/data/local/entity/Inventory.kt
@@ -27,6 +27,6 @@ data class Inventory(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val productId: Long,
     val warehouseId: Long,
-    val quantity: Int
+    val quantity: Double
 )
